+++ Epika/Application/GameRuntime/BattleEngine/BattleEngine+Reactions.swift
@@ -421,7 +421,7 @@ extension BattleEngine {
                 if result.wasCritical {
                     entryBuilder.addEffect(kind: .skillEffect,
                                            target: targetIdx,
-                                           extra: SkillEffectLogKind.magicCritical.rawValue)
+                                           extra: UInt32(SkillEffectLogKind.magicCritical.rawValue))
                 }
                 if result.guardBarrierConsumed > 0 {
                     for _ in 0..<result.guardBarrierConsumed {
@@ -442,7 +442,7 @@ extension BattleEngine {
                     entryBuilder.addEffect(kind: .magicDamage,
                                            target: targetIdx,
                                            value: UInt32(applied),
-                                           extra: UInt16(clamping: result.damage))
+                                           extra: UInt32(clamping: result.damage))
                 }
                 if absorbed > 0 {
                     entryBuilder.addEffect(kind: .healAbsorb, target: attackerIdx, value: UInt32(absorbed))
@@ -479,7 +479,7 @@ extension BattleEngine {
                 entryBuilder.addEffect(kind: .breathDamage,
                                        target: targetIdx,
                                        value: UInt32(applied),
-                                       extra: UInt16(clamping: result.damage))
+                                       extra: UInt32(clamping: result.damage))
             }
             attackResult = AttackResult(attacker: attackerCopy,
                                         defender: targetCopy,
@@ -560,7 +560,7 @@ extension BattleEngine {
                 entryBuilder.addEffect(kind: .statusRampage,
                                        target: targetIdx,
                                        value: UInt32(applied),
-                                       extra: UInt16(clamping: rawDamage))
+                                       extra: UInt32(clamping: rawDamage))
             }
 
             if !hasStatus(tag: statusTagConfusion, in: defender, state: state) {
